move test helpers back to relevant files
