+++ ui/tests/integration/components/auth/page/mfa-test.js
@@ -6,12 +6,137 @@
 import { module, test } from 'qunit';
 import { setupRenderingTest } from 'ember-qunit';
 import { setupMirage } from 'ember-cli-mirage/test-support';
-import { setupTotpMfaResponse } from 'vault/tests/helpers/mfa/mfa-helpers';
-import { mfaTests, setupTestContext } from './test-helper';
+import { constraintId, setupTotpMfaResponse } from 'vault/tests/helpers/mfa/mfa-helpers';
+import setupTestContext from './setup-test-context';
 import { ERROR_JWT_LOGIN } from 'vault/components/auth/form/oidc-jwt';
 import { overrideResponse } from 'vault/tests/helpers/stubs';
 import sinon from 'sinon';
-import { windowStub } from 'vault/tests/helpers/oidc-window-stub';
+import { triggerMessageEvent, windowStub } from 'vault/tests/helpers/oidc-window-stub';
+import { AUTH_FORM } from 'vault/tests/helpers/auth/auth-form-selectors';
+import { GENERAL } from 'vault/tests/helpers/general-selectors';
+import { MFA_SELECTORS } from 'vault/tests/helpers/mfa/mfa-selectors';
+import { fillInLoginFields } from 'vault/tests/helpers/auth/auth-helpers';
+import { click, fillIn, waitFor } from '@ember/test-helpers';
+
+const mfaTests = (test) => {
+  test('it displays mfa requirement for default paths', async function (assert) {
+    const loginKeys = Object.keys(this.loginData);
+    assert.expect(3 + loginKeys.length);
+    this.stubRequests();
+    await this.renderComponent();
+
+    // Fill in login form
+    await fillIn(AUTH_FORM.selectMethod, this.authType);
+    await fillInLoginFields(this.loginData);
+
+    if (this.authType === 'oidc') {
+      // fires "message" event which methods that rely on popup windows wait for
+      // pass mount path which is used to set :mount param in the callback url => /auth/:mount/oidc/callback
+      triggerMessageEvent(this.path);
+    }
+
+    await click(GENERAL.submitButton);
+    await waitFor(MFA_SELECTORS.mfaForm);
+    assert
+      .dom(MFA_SELECTORS.mfaForm)
+      .hasText(
+        'Back Multi-factor authentication is enabled for your account. Enter your authentication code to log in. TOTP passcode Verify'
+      );
+    await click(GENERAL.backButton);
+    assert.dom(AUTH_FORM.form).exists('clicking back returns to auth form');
+    assert.dom(AUTH_FORM.selectMethod).hasValue(this.authType, 'preserves method type on back');
+    for (const field of loginKeys) {
+      assert.dom(GENERAL.inputByAttr(field)).hasValue('', `${field} input clears on back`);
+    }
+  });
+
+  test('it displays mfa requirement for custom paths', async function (assert) {
+    this.path = `${this.authType}-custom`;
+    const loginKeys = Object.keys(this.loginData);
+    assert.expect(3 + loginKeys.length);
+    this.stubRequests();
+    await this.renderComponent();
+
+    // Fill in login form
+    await fillIn(AUTH_FORM.selectMethod, this.authType);
+    // Toggle more options to input a custom mount path
+    await fillInLoginFields({ ...this.loginData, path: this.path }, { toggleOptions: true });
+
+    if (this.authType === 'oidc') {
+      // fires "message" event which methods that rely on popup windows wait for
+      triggerMessageEvent(this.path);
+    }
+
+    await click(GENERAL.submitButton);
+    await waitFor(MFA_SELECTORS.mfaForm);
+    assert
+      .dom(MFA_SELECTORS.mfaForm)
+      .hasText(
+        'Back Multi-factor authentication is enabled for your account. Enter your authentication code to log in. TOTP passcode Verify'
+      );
+    await click(GENERAL.backButton);
+    assert.dom(AUTH_FORM.form).exists('clicking back returns to auth form');
+    assert.dom(AUTH_FORM.selectMethod).hasValue(this.authType, 'preserves method type on back');
+    for (const field of loginKeys) {
+      assert.dom(GENERAL.inputByAttr(field)).hasValue('', `${field} input clears on back`);
+    }
+  });
+
+  test('it submits mfa requirement for default paths', async function (assert) {
+    assert.expect(2);
+    this.stubRequests();
+    await this.renderComponent();
+
+    const expectedOtp = '12345';
+    this.server.post('/sys/mfa/validate', async (_, req) => {
+      const [actualOtp] = JSON.parse(req.requestBody).mfa_payload[constraintId];
+      assert.true(true, 'it makes request to mfa validate endpoint');
+      assert.strictEqual(actualOtp, expectedOtp, 'payload contains otp');
+    });
+
+    // Fill in login form
+    await fillIn(AUTH_FORM.selectMethod, this.authType);
+    await fillInLoginFields(this.loginData);
+
+    if (this.authType === 'oidc') {
+      // fires "message" event which methods that rely on popup windows wait for
+      triggerMessageEvent(this.path);
+    }
+
+    await click(GENERAL.submitButton);
+    await waitFor(MFA_SELECTORS.mfaForm);
+    await fillIn(MFA_SELECTORS.passcode(0), expectedOtp);
+    await click(MFA_SELECTORS.validate);
+  });
+
+  test('it submits mfa requirement for custom paths', async function (assert) {
+    assert.expect(2);
+    this.path = `${this.authType}-custom`;
+    this.stubRequests();
+    await this.renderComponent();
+
+    const expectedOtp = '12345';
+    this.server.post('/sys/mfa/validate', async (_, req) => {
+      const [actualOtp] = JSON.parse(req.requestBody).mfa_payload[constraintId];
+      assert.true(true, 'it makes request to mfa validate endpoint');
+      assert.strictEqual(actualOtp, expectedOtp, 'payload contains otp');
+    });
+
+    // Fill in login form
+    await fillIn(AUTH_FORM.selectMethod, this.authType);
+    // Toggle more options to input a custom mount path
+    await fillInLoginFields({ ...this.loginData, path: this.path }, { toggleOptions: true });
+
+    if (this.authType === 'oidc') {
+      triggerMessageEvent(this.path);
+    }
+
+    await click(GENERAL.submitButton);
+    await waitFor(MFA_SELECTORS.mfaForm);
+    await fillIn(MFA_SELECTORS.passcode(0), expectedOtp);
+    await click(MFA_SELECTORS.validate);
+  });
+};
 
 module('Integration | Component | auth | page | mfa', function (hooks) {
   setupRenderingTest(hooks);
